Update tooltip.js test
Update to the tooltip.js test to capture whether an aria-label was added that preserves the text of the title, and to test whether an original aria-label attribute is preserved, rather than overwritten by the tooltip's title.

This is to address cases where screen readers are not capturing the aria-describedby attribute that is added when the tooltip is triggered.  This should also avoid a race condition between the screen reader and the appearance of the tooltip.